debug: implement assert macro
TODO: replace some of the die calls with this? (to allow for smaller
code size, when running without debugs).

Alternative: Introduce macro-setting which still keeps die calls, but
replaces it with parameter-less, so that the strings are discarded,
and thus not included in the binary, for space-saving, while still
keeping the error-checking.