debug: Disable IRQ when we enter function of death
TODO: We should still try to end up in one of the re-defineable
handlers, or something, so that the program can define how a fatal
error should be handled (it might need to close some relays, or such).
Except when debugging, when it might be more useful to just BKPT (as
now).