debug: Refactor assertions + add assert_always + static_assert
On GCC (which we are currently bound to use, but I'm trying to use
defines, so we can be somewhat portable to other C compilers), we also
generate an error at compile-time (using a function with error
attribute) if it is possible to evaluate the assertion at this point.
This is useful for inlined functions with assertions.

Also we can now write static_assert for _Static_assert

TODO: consider having a function like build_assert/static_warning as
an alternative to static_assert which evaluates the assertion at
compile-time if it can, but simply ignoring it if it is not possible,
while adding nothing to runtime-code.