debug: rename print_hex and println_hex to print_u32_hex and println_u32_hex
This should make it easier to expand in the future with functions that
print different sizes, and for instance print_i32_dec and
print_u32_dec, etc. etc.

Further on these could be adapted to a printf-like function (should
probably not be printing directly, but working against a string buffer
in that case, or perhaps a rope. To do the latter without a heap would
require using a pseudo-continuation passing style recursive definition
of printf (pass the rest of the cmd string), or a heap)