nRF24L01: Move all static options into its own struct
This way we can save RAM by having all needed info about the device in
flash, by declaring it static const.

If the conf needs to be in RAM for speed reasons it is very easy to
declare it as so as well.

However one major con of this approach, still, is that most accesses
in the nRF24L01 functions are now doubly indirect.