nRF24L01-send-test: semi-working (debugging, WIP)
Seems we have to set the NSS pin as an AF output/input (have only
tried the former) for the SPI peripheral to work. So change to use A3
for CSN, as we want that to be controlled like a regular old GPIO.
Consider setting SSOE = 1 ...

Toggle pin 15 to show runtime of spi_setup and nRF24L01_init. Put in
characteristic delays around to make it easier to find in logic
analyzer output.