nRF24L01-test: refactor all port/pin mappings out to header file
IRQ is noticeably messier because we need to use so many disparate
peripherals to set up the external interrupt. TODO: helper function
for setting up external interrupts reducing this to 2-3 defines (e.g.
it takes a GPIO_PortPin, finds out the relevant settings from it,
landing us at 2 defines: The GPIO_PortPin and the name of the
interrupt routine.