pcecon-receiver-simple first commit
This is a less abstracted implementation (more is hard-coded) of a
pcecon receiver that aims at increased speed (lower interrupt
latency), at the expense of flexibility. It does not use the
pceCon_client library.

For instance now the GPIO pins have to be in order and on the same
port (since we manipulate the pins in paralell with ODR), and the
enable_not pin is not emulated entirely correctly (but since the PCE
seems to be using this pin mostly like a clock tick for auto-fire
etc., this does not matter much).